# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "오뚜기 파워세일즈 API",
    "description": "B2B 영업사원 실적 조회 및 목표 관리 시스템 API",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:49275",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "Bearer Authentication": []
    }
  ],
  "paths": {
...
    "/api/v1/admin/education/posts/{postId}": {
      "get": {
        "tags": [
          "admin-education-controller"
        ],
        "operationId": "getPostDetail",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseEducationPostDetailResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "admin-education-controller"
        ],
        "operationId": "updatePost",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "title",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "content",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "files",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "binary"
              }
            }
          },
          {
            "name": "keepFileKeys",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseEducationMutationResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin-education-controller"
        ],
        "operationId": "deletePost",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/mobile/auth/verify-password": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "verifyPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mobile/auth/refresh": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTokenResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/mobile/auth/login": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseLoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mobile/auth/gps-consent": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "recordGpsConsent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GpsConsentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseGpsConsentRecordResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mobile/auth/change-password": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "changePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/notices": {
      "get": {
        "tags": [
          "admin-notice-controller"
        ],
        "operationId": "getPosts",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseNoticePostListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin-notice-controller"
        ],
        "operationId": "createNotice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoticeCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseNoticeMutationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/education/posts": {
      "get": {
        "tags": [
          "admin-education-controller"
        ],
        "operationId": "getPosts_1",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminEducationListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin-education-controller"
        ],
        "operationId": "createPost",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "content",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "files",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseEducationMutationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auth/refresh": {
      "post": {
        "tags": [
          "admin-auth-controller"
        ],
        "operationId": "refresh_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auth/login": {
      "post": {
        "tags": [
          "admin-auth-controller"
        ],
        "operationId": "login_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminLoginResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/mobile/notices": {
      "get": {
        "tags": [
          "notice-controller"
        ],
        "operationId": "getPosts_2",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseNoticePostListResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/mobile/education/posts": {
      "get": {
        "tags": [
          "education-controller"
        ],
        "operationId": "getPosts_3",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseEducationPostListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mobile/education/posts/{postId}": {
      "get": {
        "tags": [
          "education-controller"
        ],
        "operationId": "getPostDetail_1",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseEducationPostDetailResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "ApiResponseObject": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
...
      "ApiResponseNoticeMutationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/NoticeMutationResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
...
      "ApiResponseEducationMutationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/EducationMutationResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
...
      "EducationMutationResponse": {
        "type": "object",
        "properties": {
          "eduId": {
            "type": "string"
          },
          "eduTitle": {
            "type": "string"
          },
          "eduContent": {
            "type": "string"
          },
          "eduCode": {
            "type": "string"
          },
          "eduCodeNm": {
            "type": "string"
          },
          "employeeId": {
            "type": "integer",
            "format": "int64"
          },
          "instDate": {
            "type": "string"
          },
          "updDate": {
            "type": "string"
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttachmentInfo"
            }
          }
        },
        "required": [
          "attachments",
          "eduCode",
          "eduCodeNm",
          "eduContent",
          "eduId",
          "eduTitle",
          "instDate"
        ]
      },
...
      "VerifyPasswordRequest": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          }
        },
        "required": [
          "password"
        ]
      },
      "RefreshTokenRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "ApiResponseTokenResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/TokenResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "accessToken",
          "expiresIn",
          "refreshToken"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "employeeCode": {
            "type": "string",
            "pattern": "^\\d{8}$"
          },
          "password": {
            "type": "string",
            "maxLength": 2147483647,
            "minLength": 4
          },
          "deviceId": {
            "type": "string"
          }
        },
        "required": [
          "employeeCode",
          "password"
        ]
      },
      "ApiResponseLoginResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/LoginResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserInfo"
          },
          "token": {
            "$ref": "#/components/schemas/TokenInfo"
          },
          "requiresPasswordChange": {
            "type": "boolean"
          },
          "requiresGpsConsent": {
            "type": "boolean"
          }
        },
        "required": [
          "requiresGpsConsent",
          "requiresPasswordChange",
          "token",
          "user"
        ]
      },
...
      "GpsConsentRequest": {
        "type": "object",
        "properties": {
          "agreementNumber": {
            "type": "string"
          }
        }
      },
      "ApiResponseGpsConsentRecordResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/GpsConsentRecordResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
      "GpsConsentRecordResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "accessToken",
          "expiresIn"
        ]
      },
      "ChangePasswordRequest": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string",
            "maxLength": 2147483647,
            "minLength": 4
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
...
      "NoticeCreateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          },
          "category": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "branch": {
            "type": "string"
          },
          "branchCode": {
            "type": "string"
          }
        },
        "required": [
          "category",
          "content",
          "title"
        ]
      },
...
      "ApiResponseAdminLoginResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/AdminLoginResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
...
      "ApiResponseNoticePostListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/NoticePostListResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
...
      "ApiResponseEducationPostListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/EducationPostListResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
...
      "ApiResponseEducationPostDetailResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/EducationPostDetailResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
...
      "EducationPostDetailResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "categoryName": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EducationAttachmentResponse"
            }
          }
        },
        "required": [
          "attachments",
          "category",
          "categoryName",
          "content",
          "createdAt",
          "id",
          "images",
          "title"
        ]
      },
...
      "ApiResponseAdminEducationListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/AdminEducationListResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
...
    },
    "securitySchemes": {
      "Bearer Authentication": {
        "type": "http",
        "name": "Bearer Authentication",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}